Uber commit: fragletc(+tests) + embedded envelopes
By embedding the existing envelopes into our tools, we'll always ship
something that works. But we also need the flexibility to BYOE
(bring-your-own-envelop) as well, we should be sure that every envelope
that we embed is well tested.

The envelopes_test\ is **not** what we want, it's the reason that
motivated some of the above changes.

Aside: This CLI test framework was worth every penny of investment.